pakoreguota medianos skaiciavimo funkcija

+++ vektorius.cpp
@@ -21,8 +21,8 @@ double skaiciuotiMediana(vector<int> paz) {
     if (paz.empty()) return 0.0;
     sort(paz.begin(), paz.end());
     size_t n = paz.size();
-    if (n % 2 == 0) return (static_cast<double>(paz[n / 2 - 1]) + static_cast<double>(paz[n / 2] / 2.0));
-    else return static_cast<double>(paz[n / 2]);
+    if (n % 2 == 0) return paz[n / 2 - 1] + paz[n / 2] / 2.0;
+    else return paz[n / 2];
 }
 
 void apskaiciuotiPagalMetoda(Studentas& st, int metodas) {
